refactor: move code place
