+++ frontend/src/components/InstanceForm/index.ts
@@ -4,6 +4,8 @@ import SpannerHostInput from "./SpannerHostInput.vue";
 import SpannerCredentialInput from "./SpannerCredentialInput.vue";
 import OracleSIDAndServiceNameInput from "./OracleSIDAndServiceNameInput.vue";
 
+import InstanceForm from "./InstanceForm.vue";
+
 export {
   SslCertificateForm,
   SshConnectionForm,
@@ -11,3 +13,5 @@ export {
   SpannerCredentialInput,
   OracleSIDAndServiceNameInput,
 };
+
+export default InstanceForm;
